adjust project structure
