+++ src/app/dashboard/layout.tsx
@@ -5,8 +5,8 @@ import Container from "@mui/material/Container";
 import GlobalStyles from "@mui/material/GlobalStyles";
 
 import { AuthGuard } from "@/components/auth/auth-guard";
-import { MainNav } from "@/components/dashboard/layout/main-nav";
-import { SideNav } from "@/components/dashboard/layout/side-nav";
+import { MainNav } from "@/components/dashboard/layout/MainNav";
+import { SideNav } from "@/components/dashboard/layout/SideNav";
 import { config } from "@/config";
 
 export const metadata = {
